Fake HTTP server

+++ Client/app/Config.hpp
@@ -7,7 +7,7 @@
 namespace app { namespace config {
 
   extern const std::string serverIp;
-  extern const short serverPort;
+  extern const int serverPort;
 
 } }
