+++ Client/src/Config.cpp
@@ -3,6 +3,6 @@
 namespace app { namespace config {
 
   const std::string serverIp = "0.0.0.0";
-  const short serverPort = 8889;
+  const int serverPort = 8889;
 
 } }
