+++ Common/net/socket/Udp.hpp
@@ -12,7 +12,7 @@ namespace common { namespace net { namespace socket {
       typedef asio::ip::udp::endpoint EndPoint;
 
     public:
-      Udp(const std::string &/* ip */, const short &/* port */);
+      Udp(const std::string &/* ip */, const int &/* port */);
       ~Udp();
 
     public:
@@ -37,7 +37,7 @@ namespace common { namespace net { namespace socket {
         const std::string &s,
         EndPoint &/* endpoint */
         );
-      EndPoint resolver(const std::string &from_adr, const short port);
+      EndPoint resolver(const std::string &from_adr, const int port);
 
     private:
       asio::ip::udp::socket m_socket;
